scripts/python/requirements.txt
functions-framework
google-cloud-bigtable